requirements.txt
Django==2.0.6
django-filter==1.1.0
djangorestframework==3.8.2
Markdown==2.6.11
pytz==2018.4
